Desktop/app domicilios/flask-icfes/requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
cryptography==41.0.3
python-dotenv==1.0.0
supabase==2.0.3
WTForms==3.1.0

# === TESTING ===
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-flask==1.2.0
pytest-mock==3.11.1

# === CODE QUALITY ===
flake8==6.1.0
black==23.9.1
bandit==1.7.5
safety==2.3.5
